refactor(components): add missing components to index export
