+++ src/renderer/src/components/index.ts
@@ -1,7 +1,12 @@
-export * from "./Stack";
 export * from "./Button";
-export * from "./TextInput";
-export * from "./Drawer";
 export * from "./ButtonLink";
-export * from "./Select";
+export * from "./ConfirmationModal";
+export * from "./DatePicker";
+export * from "./Drawer";
 export * from "./Modal";
+export * from "./RenderMarkdown";
+export * from "./Select";
+export * from "./Stack";
+export * from "./Tag";
+export * from "./TextInput";
+export * from "./VerticalButtonGroup";
